Versi 1.05 (Minggu12)
PHP-SMK Minggu 12

+++ Restoran/kategori/delete.php
@@ -1,11 +1,11 @@
 <?php
 
-require_once "../function.php";
+if (isset($_GET['id'])) {
+    $id = $_GET['id'];
 
-// $id = 3;
+    $sql = "DELETE FROM tblkategori WHERE idkategori=$id";
 
-$sql = "DELETE FROM tblkategori WHERE idkategori = $id";
+    $db->runSQL($sql);
 
-$result = mysqli_query($koneksi, $sql);
-
-header("location:http://localhost/php-smk/Restoran/kategori/select.php");
+    header("location: ?f=kategori&m=select");
+}
